Changed the query name for fetching characters belonging to an account. Added the ability for BaseSocket and ClientSocket to Send() a PEPacket object instead of pre-serialised data only Added ability to disconnect the console client on failed login. Reworked the login status handling Added row count to DBIResultSet (current updated by DBIMySqlConnection) Added the fcserverprotocol.h which will contain inter-service comms packet definitions   - Added the __FCSPKT_CLIENT_DISCONNECT message Added the Account object to the Auth Service to manage logged in accounts Completed the Auth Service's handling of accounts logging in as well as disconnecting (forced logout) Added the ability for the router to notify its connected services of client disconnects Added the start of the Eventing system for the Game World Service Defined Player and InGameIPAddress objects for use in the world (still prototype phase)

+++ trunk/common/database/mysql/DBIMySqlConnection.cpp
@@ -67,7 +67,7 @@ DBIResults* DBIMySqlConnection::Execute(FCDBJob job)
 
   bool bResult = true;
   string query = job.GetQuery();
-  int nResult = mysql_real_query(m_conn, query.c_str(), query.length());
+  int nResult = mysql_real_query(m_conn, query.c_str(), (unsigned long)query.length());
   DBIResults* pResults = new DBIResults;
   DBIMySqlResultSet* pResultSet = NULL;
